backend/back_AI_opinion_polling/requirements.txt
# Django & REST
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-allauth==65.12.1
dj-rest-auth==7.0.1

# Database (PostgreSQL)
psycopg2-binary==2.9.10

# Celery & Redis
celery==5.4.0
redis==5.1.1

# Utilities
python-dotenv==1.2.1
loguru==0.7.3
requests==2.32.5
cryptography==46.0.3
openai==2.6.1
textblob==0.19.0
pytest==8.4.2
pytest-django==4.11.1
numpy==2.3.5
pandas==2.3.3
scipy==1.16.3
tiktoken==0.12.0
scikit-learn==1.7.2
drf-spectacular==0.29.0
openpyxl==3.1.5


# Web server
gunicorn==23.0.0
